catch file handler error

+++ Server/NatureWatchServer/src/main/java/at/jku/se/database/FileHandler.java
@@ -12,25 +12,39 @@ public class FileHandler {
 
 	public void putFile(String name, byte[] data) {
 
+		if (name == null) {
+			System.out.println("Filename is null!");
+			return;
+		}
 		try {
-    
-	    Path file = Paths.get(this.directory + name);
-	    Files.write(file, data);
-	    
+
+			Path file = Paths.get(this.directory + name);
+			Files.write(file, data);
+
 		} catch (IOException e1) {
 			e1.printStackTrace();
 		}
 	}
-	
+
 	public boolean deleteFile(String name) {
+
+		if (name == null) {
+			System.out.println("Filename is null!");
+			return true;
+		}
 
 		File file = new File(this.directory + name);
 		return file.delete();
-		
+
 	}
 
 	public byte[] getFile(String name) {
 		
+		if (name == null) {
+			System.out.println("Filename is null!");
+			return null;
+		}
+
 		try {
 			File file = new File(this.directory + name);
 			return Files.readAllBytes(file.toPath());
